backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
google-genai
PyJWT
gTTS
passlib[bcrypt]
python-multipart
duckduckgo-search